CNS-870: move IPRPC functionality to different file and divide to functions
